// client-rmq/src/position/dto/create-position.dto.ts
/* eslint-disable prettier/prettier */
export class CreatePositionDto {
    deviceId: string;
    header: string;
    dateTime: string;
    dateSystem: string;
    latitude: string;
    longitude: string;
    inFence: string;
    ignition: string;
    direction: string;
    speed: string;
    bateryPercent: string;
    voltage: string;
    odometer: string;
    satellite: string;
    signal: string;
    out1: boolean;
    out2: boolean;
    cmd: string;
}